requirements.txt
altgraph==0.17.4
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pystray==0.19.5
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==75.6.0
six==1.17.0
watchdog==6.0.0
winotify==1.1.0
